Use Trilinos_PACKAGE_ENABLES_FILE and SPARCTrilinosPackagesList.cmake (ATDV-399)
This fixes the XL build on ats-2 for the packages used by SPARC and the
handling of STK (see ATDV-399).

This also allows the removal of the file SPARC_Trilinos_PACKAGES.cmake.  Now
that the file SPARCTrilinosPackagesEnables.cmake is getting included in the
outer ctest -S driver, there is no longer a need for this file in the repo.