Use SPARCMiniTrilinosPackagesEnables.cmake for ctest -S driver mini build (ATDV-399)
The file SPARCMiniTrilinosPackagesEnables.cmake is now getting included in the
outer ctest -S script and in the inner CMake configure.  There is no need for
the file SPARC_MiniTrilinos_PACKAGES.cmake so it is removed.

This is not needed by but was inspired by ATDV-399 to clean things up and to
be consistent.