Fixes BiCGStab algorithm so that it detects breakdown
There are three coefficient calculations in this implementation
of the BiCGStab algorithm that need to be checked to see if the
denominator is less than the sfmin for that scalar type.  These
checks have been added to the algorithm and logic has been added
to the solver manager to check for breakdown if the iteration
exits before convergence or the maximum number of iterations.
There is now a test that solves an inner-outer linear system
with diagonal matrices that checks for one type of breakdown
(the good kind).