MueLu: unit test for TestHelpers themselves
Although we use the TestHelpers in all the tests, we also should check,
if they deliver the correct result. This might be less difficult and
less important for simple matrices, such as 1D Poisson, but for sure is
crucial for more complex scenarios to come, e.g. 2x2 blcok sysmtes for
meshtying problems.