Fix the filepath when using Makefiles generator (#7376)
When using the Makefiles Generator, the magic_wrapper.py command gets called
from working directories under the base build directory and the output files
in the *.timing files were not giving the full relative path to the base build
dir.

We need to have the full relative path to the base build dir so that we can
bin the entries for each subdir under packages/ (which is close to TriBITS
packages).