MueLu: simplify input parameters for MapTransfer
The internal mechanism of limiting the used nullspace vectors to
translations only is the same for 2D and 3D problems. Hence, the
paramter to control this behavior should also not differ for 2D and 3D
problems. Now, we use the same value, "translations", to exclude
rotational modes from the MapTransfer and only consider translations,
despite the spatial dimension of the problem.