Make work with Python 3.6 (#7376)
Here, I just had to deal with some byte-array vs. string conversion issues.

I also had to deal with some class data issues.  I could not find a way to get
the code as written to work so I moved the data outside of the class
WrapperOpTimer.  But this is data is encapsulated in the module
WrapperOpTimer.py so there is no practical reduction in data protection here.