Set Trilinos_REMOVE_BUILD_STATS_TIMING_FILES_ON_FRESH_CONFIGURE=ON for cuda_9.2 PR build (#7376)
This should eliminate a corrupted *.timing file for this build (see PR #7508).

This option should likely get set for all the PR builds to make them more
robust.  But that will make it so that you only see the build stats for the
very last sest of targets (re)built in that particular testing iteration.