Framework: Bugfix for cycle detection PR configurations
The cycle-detection logic in the PR driver setenvironment module
had a minor bug that resulted in an overly aggressive 'cycle prevention'
mode.

For example, section dependencies like this would be blocked:
```
A -> B -> C -> B
```

but we also blocked this:
```
A -> B -> C
A -> B
```

This fixes the issue by popping off the sections from the 'seen list' as
we pop out of scope during the recursive module that scans through the
configurations.