Change from 'cee-rhel6' to 'cee-rhel7' (ATDV-391) and 'gcc' to 'gnu' (ATDV-408)
This is the long-overdue name changes of 'cee-rhel6' to 'cee-rhel7' (see
ATDV-391).

This also fixes a but introduced a long time agao that had 'gcc' in the names
of the 'cuda' builds instead of 'gnu'.  Well, that does not work with the
updated SPARC TPL installer that puts in 'gnu' instead of 'gcc' (see
ATDV-408).

NOTE: This commit is going to create conflicts with trilinos/Trilinos#8362.
(There is no way around it.)  I will just need to resolve those conflicts when
I merge this with 'develop'.