Sacado: Enable dynamic linking of host code using Sacado in Cuda builds with RDC.
Some Sierra tests that include user plugins that are dynamically linked
previously failed to compile in Cuda builds, reporting that the
global_sacado_cuda_memory_pool_on_device symbol was undefined. The
user plugin code is host only, and we don't enable the Sacado memory
pool in Sierra anyway. Checking whether or not the memory pool is
enabled in one additional ifdef resolves the undefined symbol for
the shared library creation.