MueLu: fake-extract zero block in block matrix
When a saddle-point system is represented by a BlockedCrsMatrix, then
the zero block often isn't an allocated matrix, but just Teuchos::null.
Still, for forming coarse levels, map information about this block is
required. The ZeroSubBlockAFactory fakes the zero-block by generating a
matrix with the appropriate maps, but no entries.

Tests will follow.

Part of #8586.

Co-authored-by: Matthias Mayr <matthias.mayr@unibw.de>
Co-authored-by: Tobias Wiesner <tobias@tawiesn.de>