Framework: Setting Intel build CMake back down to 3.10.3 for testing
The intel build fails the hello-world compiler check for the F90
compiler (mpif90) in CMake 3.17 and 3.19.  This commit is intended
to reset the compiler we're using for the Intel compiler build back
to 3.10.3 to see if it passes in the PR.  If it does, this gives us
a data point that there _is_ some kind of issue going on with the
newer version(s) of CMake and the Intel compiler.