Framework: disable color output in Python testing
The pytest module by default appears to colorize the output
from the tool. I think this may be the reason that a regex
rule like this:
```
^=+ [0-9]+ passed in
```
is failing on output that looks like this:
```
============================= 106 passed in 1.34s ==============================
```

There should be a match but if the ESC characters are being generated
and CMake / CTest isn't properly stripping that prior to regex matching
consideration then it could be the culprit.

If this does not fix the issue, we could just remove the `^` in the
REGEX and that should match the string too.