Shylu: pass non-const matrix to MueLu
MueLu currently only accepts non-const matrices as input to create a
hierarchy. Until this is resolved in MueLu, do a
Teuchos::rcp_const_cast() to get rid of the const qualifier prior to
handing the matrix to MueLu.

Re-enable two Laplace tests, that use MueLu as inexact subdomain solver.

Part of #8762.