Framework: Update STALE issue closer
I upped the version of the issue autocloser to 3.0.17 from 3.0.14

I increased the rate-limit threshold from 50 to 52 -- I'd like to see
if we can find a value that gets us to about 100 issues / month that we
can autoclose.

We can go higher if we set up a 'hole' in the autotester CRON on the
weekend when this runs... but given how the autotester can run for a day
or more in a single instance we could still be imposing a rate limiter
hit?