MueLu: Update region regression tests to take into account machine precision
When comparing against gold files we need to carefully check
that the values we compare have enough significant digits of
accuracy to reliably indicate a divergence from expected gold
values. To do so we added a relative error check in the
comparison script for values close to 1e0 and an absolute
error comparison for values close to machine precision.
We add two new variables in the script, one to define the
relative error to verify and another one to control when the
switch from relative error to absolute error occurs.