SEACAS: Disable tests that are failing on vortex
These SEACAS tests diff the stderr output with a known stderr gold file.  On vortex, the stderr is polluted with some jsrun output that messes up the comparisions.  It would be possible to do some grepping and other filtering to clean up the stderr, but these tests are well covered on other systems.

Should fix #8797