Fix csv field ordering
The real fix is to apply a sort to the fields before writing. I also
added the ability to write only certain fields - which is neat, but
also means that we will not have detailed data if used.

By using sort, this fix will ensure that for a given build, all timing
files should have the same field order.

CommandLineParser now scans the ENV for TRILINOS_BUILD_STATS_OUTPUT_FIELDS
Which can optionally be used to specify only a few fields to write out.

The field order can still change if new fields are added (between builds).
Or if the scripts are changed. It would probably be productive to
assume:
  CSV field orderings for a given build will be consistent
  Field ordering between builds can change.

Given this is likely used in CI testing, where builds are intentionally
*kept* - perhaps the answer is to take the current csv fields, and set
TRILINOS_BUILD_STATS_OUTPUT_FIELDS=<everything we have now> (in
whatever order you want).

Then, if we add fields later, there would still need to be logic to
ensure that a reliable merge of old + new data can happen (so that
problem never really goes away).