ats2: Print/pass jsrun args with double quotes instead of single quotes (#8474)
Using "<arg>" instead of '<arg>' seems more correct and consistent with how
CTest passes arguments.

This is to help with debugging #8474.  This does technicall change how
arguments are passed into jsrun.  Not sure if this is the right thing to do or
not since passing in as "<something>" allows for evaluation of the internals
while '<something>' does not in bash.