MueLu: Draft tutorial using ReadTheDocs
This commit is the combined work of Matthias Mayr, Graham Harper, and
Max Bahnmueller.  With their consent, I am adding a sign-off.

Signed-off-by: Jonathan Hu <jhu@sandia.gov>

message from first commit:
"This is just a draft to get Max Bahnmueller started on porting the MueLu
tutorial from Latex to rst for ReadTheDocs. This is not yet properly
integrated into TriBITS, but uses a manual in-source-directory build. I
will look into this later."