Fix gather_build_stats.py and tests for Python 3.5.2 (#7376)
With Python 3, you can't sort a list of dicts without specifying a key to sort
by (using a lambda function).  Also, with Python 3 aDict.keys() does not
return a simple list but instead it returns a dict_keys object that does not
have a sort() function.  So instead, you have to call sorted(aDict.keys()) to
get a sorted list of keys.