Make tests pass independent of the order that *.timing files are found (#7376)
This production code was not impacted, just the check in the tests.  To fix
this, I just had to sort the rows on the generated and exteced CSV file text
so it does not matter what order that the *.timing files are found.

I also eliminated some duplication in the tests.