Print stats about *.timing files and handle empty keys list (#7376)
This is to try to debug a problem that is happening with the Trilinos CUDA PR
build.  In that build, the superset of keys list appears to be empty in which
case you can't get the list for the first key.

We are clearly missing a test case for when keys list would be empty but I
need to see what is happening in that CUDA PR build so I can understand what
we need to be testing for.