Ifpack2 Chebyshev: Add options for power method
"chebyshev: eigenvalue relative tolerance": allows to terminate the
power iteration based on a relative tolerance instead of a predefined
max number of iterations

"chebyshev: eigenvalue keep vectors": When true, stores the vectors
used by the power method. This can be beneficial when the matrix is
changed only marginally between setups.