PyTrilinos: Fix PyTrilinos.Teuchos build error
The Teuchos::ParameterList was updated to utilize a
Teuchos::ParameterListModifier class, which was not wrapped, so SWIG
did not know how to handle it.  The temporary fix implemented here is
that all methods (including a constructor) that reference
ParameterListModifier are now ignored.  A better fix might be to wrap
the ParameterListModifier and not ignore those methods.