MueLu: interpolation order input changes for GeometricInterpolation
Geometric interpolation order should be specified in the structured
aggregation (coarsening order). With these changes the user should
input the interpolation order once in the structured aggregation
factory. The geometric interpolation factory obtains this parameter from
the structured aggregation factory. This requires to update the .xml
files used for the structured algorithm. On the plus side, the new
approach is less error prone since the same parameter has to be
specified only once.