Make gather_build_stats.py silent by default but add -v option (#7376)
This makes the 'gather-build-stat' target completely quiet.  This responds to
feedback from @jjellio that the make command was a bit verbose (and I agree).

But I added the -v option and I used that in the TrilinosBuildStats_Results
test so that you can see the statistics so it will be shown on CDash.