Adds more un-preconditioned Tpetra tests to Belos
This commit translates more of the Epetra-based tests to Tpetra
linear algebra.

1)  The Tpetra solver factory test was extended to include all of
the linear solvers that are accessible through the Belos linear
solver factory.  This resulted in the detection of a bug in the
test that was causing solver failures, which has been addressed.

2)  Adds diagonal test to MINRES and TFQMR solvers for Epetra.
The extension of the Tpetra solver factory test illustrated a
problem in the MINRES solver for diagonal/identity matrices,
which has been addressed in this commit.

3)  Adds BiCGStab, RCG and TFQMR test to Tpetra-based linear
algebra.  Testing for these solvers has been absent before the
extention to the Tpetra solver factory and the addition of these
tests.