Fixes MINRES solver to perform correctly for right-hand preconditioning
The MINRES solver had been written to work with left preconditioning,
but doesn't work with right-preconditioning or preconditioning on both
sides.  That has been corrected with this commit, as well as a test
including preconditioning for Epetra.

Tpetra tests for all solvers, that includes preconditioners, will be
added in a future commit.