Adding support for 'long double' scalar type when ETI is turned on.
To turn on the code, the following configure option must be added to
Trilinos:

 -D Trilinos_ENABLE_LONG_DOUBLE:BOOL=ON \

Adding support required adding LAPACK routines to Teuchos analogous to
those added to support float128 (QUADMATH).  However, the bulk of these
routines are not implemented right now and just throw when called.
These can be implemented later, if needed.

A consequence of the missing routines is that not all solvers will work
with the 'long double' scalartype.  BelosBlockGmresSolver will work,
but BelosBlockCGSolver will not, as a concrete example.