Tpetra: fix MiniFE example vectors for >1 rank
generate_miniFE_vector was only setting values owned by rank 0. This
is used to generate the RHS (b) in PerformanceCGSolve, so this caused
different initial residuals and effectively meant the linear problem got
easier as ranks increased. Now the vector is always the same for a given grid
size, as it should be.

Also got rid of the x, r, Ap vector initialization. r and Ap are
actually completely overwritten by apply/update before they are read, and
x is a freshly constructed (0-initialized) vector. This saves a
significant amount of time and makes the StackedTimer total more
representative of SpMV/BLAS that it's supposed to be measuring.