atdm: sems-rhel7: Change to what is actually being tested (#7376)
In commit 36b53f6 jmgate tried to update this to clang-10 but that does not
automatically update what is actaully being run on jenkins and displaysed on
CDash and therefore listed builds in this file along does not make them
supported builds.  Someone needs to actually add the driver scripts for these
builds under cmake/ctest/drivers/atdm/sems-rhel7/drivers/ and update the
Jenkins jobs, and triage any new failing Trilinos tests.

Making this change allows one to run 'ctest-s-local-test-driver.sh all'
properly (as I was trying to do with #7376).