Removes unnecessary BLAS/LAPACK classes and exception handling
This is just a small clean up of the solvers in Belos to remove
the unnecessary inclusion of BLAS/LAPACK classes and exception handling
that likely resulted from copying old solvers to create new ones.
The impending refactor of the code to address the dense matrix
object used internally by most solvers necessitates a clear understanding
of the dependencies of each solver on BLAS/LAPACK methods.