Tempus:  Add support for specifying the adjoint through a 2nd ME
Very few Trilinos solvers/preconditioners support solving transposed
linear systems, therefore we often have to resort to explicitly
transposing the matrix or assembling a transposed matrix directly for
adjoint sensitivities.  To better support this case, I extended the
Tempus adjoint integrator implementations to accept an additional
"adjoint" model evaluator for computing the adjoint W/W_op.  This is
necessary because there is currently no adjoint outArg that can be
requested.  I also included an adapter for the case of an implicitly
computed transpose, allowing for a single implementation.  I also
included a prototype Tpetra implementation that explicitly forms the
transpose, but it isn't tested.  Finally I modified the SinCosModel to
provide an explicit adjoint ME, and modified one of the tests to use it.

In a subsequent commit, the Thyra and Tpetra files will be moved to their
proper locations.