SEACAS: Aprepro -- remove bad stream deletion
It is not always correct to delete infoStream when it is not set to default. It could be set to the same stream as warnStream and errorStream or it could be a std stream.  For now, remove the deletion and flush the stream.  Will determine how to know whether to delete soon...