MueLu: fix misleading indentation in unit test
A mismatch in indentationn and curly braces embracing the body of an
if-statement is resolved. Although this fix changes the input map for
a unit test of the MapTransferFactory, this new version now actually
resembles the originally intended test design. Due to the test
construction, further changes to the test are not required, as the test
result is extracted from the input map. Hence, the change in the input
automatically adapts the expected outcome.

Closes #9484.