epetra: Fix size of EPETRA_OBJECT_PTR on Windows
On Win64 a `long int` is 32 bit, so not large enough to fit a pointer,
which is the point of this object. Ideally we would use some stdint,
like intptr_t here, but I assume there's all sorts of legacy reasons
that that isn't being used, so simply switch this to `long long int`
on all platforms, which is 64 bit everywhere.