teuchos: Fix syntax for explicit template instantiation
The standard syntax (and only syntax for this accepted by GCC)
is `template class <instantiation>` rather than simply
`template <instantitation>`. Since this is guarded by WIN32,
it is possible MSVC accepts the latter as an extension (I
did not check this). However, to make it build everywhere,
use the standard syntax.