Tempus: add default flag to delay initialization
Edge case: when app is only using its own TSCS in a composite, TSC
fails initialization because the app has not had the chance to add its
TSCS to the compose vector yet. So this `runInitialize` (default: true)
allow the app to delay the initialization until it has populated the
composite vector and explicitly calls the initialization.