Made TensorBasis default implementation support OPERATOR_VALUE and OPERATOR_GRAD.  (TODO: delete implementations in subclasses that handle these for the five-argument getValues().)
Revised quadrature in BasisEquivalenceTests to deal correctly with extrusions.

Right now, the new basis equivalence test fails with an exception in getCubatureImpl(), where there's an assertion that there are either two or three direct cubatures.  (And the copy logic only handles those two cases.)  So we also need (probably) to revise this to support arbitrary dimensions.