Belos: fix issue with MultiVector modify flags
Diagonal matrix apply (for testing) was using Tpetra's MV::getLocalView<device_type>
but this was incorrectly setting the "modified on host" flag when the device is
Cuda/CudaUVMSpace. Using MV::getLocalViewDevice instead correctly sets the
"modified on device" flag. This fixes 2 Belos test failures for when UVM
enabled in Tpetra, but doesn't fix the behavior of
getLocalView<device_type>.