MueLu: convert nullspace dim error to warning
In many cases, the dimension of the nullspace is >= the number of
unknonws per fine level mesh node, e.g. for solid mechanics of
Navier-Stokes. However, there are scenarios, where the dimension of the
nullspace can be smaller than the number of unknowns per fine level
node, e.g. for beam finite elements (I'll mention a particular example
below for illustration.) To allow MueLu to be used for such structural
models, we must be able to run this wihtout throwing an error.

As it still makes sense to perform this check for standard cases, the
check of the nullspace dimension stays in place, but the user can
suppress it by providing a particular option in the HierarhcyManager.

Here's the beam FEM example: For torsion-free Kirchhoff-Love beam
formulations, one ends up with 6 unkonwns per node (3 displacements, 3
"changes" in the tangent vector), but only 5 nullspace vectors (3
translations, 2 rotations around axes orthogonal to the beam
centerline). A third rotation is missing since these beams do not
possess a rigid body mode for the rotation around their own
axis/centerline.

By default, we still perform the check to help users to get their
nullspace correct.