Finish transition from sems-xyz to sems-archive-xyz (SEHELPD-2963)
This completes the transtion started in the commit:

18c4a2015d4 "change calls for sems-[TPLname] to sems-archive-[TPLname]."
Author: Henry Swantner <HRSwant@sandia.gov>
Date:   Mon Nov 22 16:52:04 2021 -0700

I tested this on a CEE RHEL7 machine manually and it seems to work correctly.

Note that I had to put in special logic to load source the script:

  /projects/sems/modulefiles/utils/sems-archive-modules-init.sh

if it has not already been sourced.  This logic was the most robust test could
figure out.