MueLu: region MG new adapter using composite data structures
Adding a new interface to solver the region problem, this
interface accepts A, X and B in composite format and still
needs the hierarchy in region format. With such an interface
hooking the preconditioner with Belos should be a lot easier
although we need to be careful how the composite vectors are
transfered to region format.

Next step will be adding a CG solver using composite vectors
and compostie A but preconditioned with region MG.