MueLu: region MG driver update and CG fix
Updating the region MG driver to allow switch between region solver,
composite Richardson and composite CG at runtime. This makes it
easier to debug code and also just makes more sense in the long run
as we will probably want to switch between various solvers depending
on the problem we try to solve.

Adding some code fixes for CG, it now converges on one of our regression
tests which is a good sign but it is still slower than the Richardson
iteration which probably points to more issues.