Enables code for multi-threaded block RHS solves based on Trilinos_ENABLE_OpenMP
This code has been placed behind a non-standard ifdef KLU_ENABLE_OMP that was only known
and used by Xyce when configuring the hybrid-hybrid ShyLU solver.  In an effort to more
easily enable and evaluate this solver for customer circuits, a transition is being made
to enabled this code using a more standardized approach, which uses Trilinos_ENABLE_OpenMP.
The TrilinosSS package now has a package-specific ifdef that is enabled whenever Trilinos
if built with OpenMP and enables this code.  It will allow for easier integration of the
hybrid-hybrid ShyLU solver.

The motivation for doing this is supported by the discussion in the Jira issue:

https://sems-atlassian-son.sandia.gov/jira/browse/TRILINOS-203