Implements the achievedTol interface for the GmresPolySolMgr
A user request has been received for the Gmres polynomial preconditioned
solver to return the achieved tolerance from the outer iterative method.
This is best done through the achievedTol() method in the abstract solver
manager interface, that is utilized by most of the iterative methods.
This commit implements that function call and includes a test of that call
to ensure it works.