Change schur inverse data type
For future modifications to the inverse
approximation factory it makes sense to
change the inverse data type from a
vector to a matrix. This of course
has small drawbacks regarding performance
(one additional matrix-matrix multiplication,
instead of a matrix-vector one). But gives
greater flexibility on how an inverse is
approximated.

- Changing the datatype of the inverse from
  a vector to a matrix
- Modifiying the inverse appoximation factory
  unit testing
- Add a blocked version to the matrix constructor
  for diagonal matrices initialized by a vector
- Add some unwrapping possibilities to the
  blocked matrix multiplication