KokkosKernels/Tpetra: Restore numVecs=1 BLAS path
This patches in Kokkos Kernels PR #1354, which restores the special
path in BLAS-1 MV reduction-like kernels for when there is only one
column. This path just calls the single-column kernel, which is a simple
parallel_reduce. This path existed before but I accidentally removed it.
This also reverts #10415 since Tpetra won't need to intercept the special path
if Kokkos Kernels is taking care of it.