Sacado: Call target_include_directories() for new TriBITS (TriBITSPub/TriBITS#299)
New TriBITS using modern CMake targets no longer sets the INCLUDE_DIRECTORIES
directory property but instead passes along include dirs using
INTERFACE_INCLUDE_DIRECTORIES target properties (like a good CMake project).
This was a strange test that does not actually link to the Sacado library but
yet needs a bunch of Sacado include dirs?

We need to look into this and see how to make this work more cleanly either in
Sacado or TriBITS because this is unfortunate.