Zoltan2: Use single value for FAIL_REGULAR_EXPRESSION (TriBITSPub/TriBITS#464)
Tt turns out that these tests really did not need to pass a list of values to
FAIL_REGULAR_EXPRESSION.  Each test was really only looking for a single value
of 'FAILED' or 'BUMMER'.  And you can argue that the tests are stronger and
better by taking out the other value that is not supposed to be printed in
that test case.