Zoltan2: Properly pass in list to FAIL_REGULAR_EXPRESSION (TriBITSPub/TriBITS#464)
Before, this was passing in:

   FAIL_REGULAR_EXPRESSION "FAIL;BUMMER"

which just happened to be added as a list to the test property with the old
TriBITS.  But with the new TriBITS in TriBITSPub/TriBITS#464 which more
correctly deals with semi-colons and list arguments, this was being set as a
single property value "FAIL\\;BUMMER" which did not match the output.

Passing in the list with:

   FAIL_REGULAR_EXPRESSION FAIL BUMMER

works both with old and new TriBITS and is more logical given how CMake/CTest
is supposed to be used.  (The old TriBITS documentation was incorrect.)