Fix issue with Clang gpu compilation.
Clang does not like dynamic initialization of a class variable when
using a gpu/device target. It may be possible to address this in
other ways, namely adding an explicit initialization when the class ETI
is processed.  Given that most (every?) solver replicated this default
pattern, and all use std::cout as the default, this change removes the
explicit variable with the default variable (which was const anyway).

This potentialy may causes issues with Windows, which was already using
a workaround to capture a pointer to cout.

packages/belos/src/BelosBlockGmresSolMgr.hpp:340:37:
error: dynamic initialization is not supported for __device__,
       __constant__, __shared__, and __managed__ variables.
    static constexpr std::ostream * outputStream_default_ = &std::cout;
                                    ^                       ~~~~~~~~~~

Fixes #10535