Intrepid2: Add support for orientation on sides.
For a 3D cell, we compute the orientations of its faces and edges, not for the cell itself,
because they are not needed for prescribing continuity across elements.
For a 2D cell, we used to compute the orientations for edges, not for the cell itself.
However, when the 2D cell is the side of a 3D cell, it inherites the orientation of the 3D cell,
and it is useful, for side projections, to support the orientation of the 2d cell itself.
Similarly, a 1D side cell would need support for its orientation.
This commits provide such support. We also add orientations for 1D and 2D HVOL basis, which are used when
HDiv functions are restricted to the boundary (in the normal direction).