Only add Krino tests when tests are enabled (TriBITSPub/TriBITS#299)
With the new TriBITS that uses modern CMake targets the upstream test package
include dirs for Krino are no longer added by tribits_add_executable().  This
causes a build error in the target Akri_Unit_main.cpp.o built under the
unit_tests/ directory for the missing header file:

../packages/krino/krino/unit_tests/Akri_Unit_main.cpp:12:10: fatal error: stk_unit_test_utils/ParallelGtestOutput.hpp: No such file or directory
 #include <stk_unit_test_utils/ParallelGtestOutput.hpp>
          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The issue is that setting the INCLUDE_DIRECTORIES property at any point in a
CMakeLists.txt file results in that value being propagated to all targets
*above* where it was set and all of the subdirectories included above.  So, having:

  add_subdirectory(unit_tests)
  ...
  tribits_add_executable(...)

is identical in behavior to:

  tribits_add_executable(...)
  ...
  add_subdirectory(unit_tests)

because the INCLUDE_DIRECTORIES property set by tribits_add_executable()
impacts everything in that directory and all subdirectories.  (This is why
old-school CMake is so problematic.)

The fix is to only add the unit_tests/ directory when tests for the Krino
package are enabled.  This fix is backward compatible with the old version of
TriBITS as well obviously because a package should not be building targets for
tests unless tests are enabled for the package.