Allow working with GenConfig repos with gitdist
This allows using the tool gitdist to view and work with the GenConfig repos.

NOTE: Having this .gitdist.default file does not impact the git submodule
manipulation as defined in .gitsubmodules or
packages/framework/get_dependencies.sh at all.  All it does it provide another
easy way to view and work with these repos.