Adds getEigenEstimates() method and changes positive-definite error
The getEigenEstimates() method returns all the eigenvalues computed
by LAPACK during the condition estimate computation.  The condition
estimate code is only provided by the pseudo-block CG solver at this
time.  There are Epetra and Tpetra tests updated to call this method.

A new CGPositiveDefiniteFailure error class created to enable users
to trap iteration errors that are specifically due to the positive
definiteness checks in the CG solvers.  Any CG iteration that uses
the errors provided in the BelosCGIteration header have been changed
to throw this new error.  The error inherits from the CGIterateFailure
class to address backwards compatibility.

This commit addresses the user request in this issue:

https://github.com/trilinos/Trilinos/issues/10446