Give Trilinos own FindTPL*.cmake modules for BLAS and LAPACK (TriBITSPub/TriBITS#299)
Somehow, the Trilinos 'develop' branch got a different version of the modules
tribits/common_tools/FindTPLBLAS.cmake and FindTPLLAPACK.cmake that what is in
the TriBITS 'master' branch.  That problem with that is that if you are
building Trilinos with external TriBITS with:

  -DTrilinos_TRIBITS_DIR:STRING=TriBITS/tribits

you get different behavior for finding BLAS and LAPACK than if you are using
the versions in Trilinos.  That is not good and it has cost people real lost
time (e.g. the testing of Trilinos PR #10614 with an offline email
discussion as part of working TriBITSPub/TriBITS#299).

The fix implemented here was to:

* Copy these modules from Trilinos/cmake/tribits/common_tools/ to
  Trilinos/cmake/TPLs/.

* Point BLAS and LAPACK in Trilinos/TPLsList.cmake to FindTPL*.cmake files
  under cmake/TPLs/ instead of cmake/tribits/common_tools/.

* Make the versions of the FindTPL*.cmake under cmake/tribits/common_tpls/ on
  the Trilinos develop branch to be identical to what is in the currently
  snapshotted version of TriBITS in the merged-in 'tribits_github_snapshot'
  branch.

The current version of TriBITS 'master' snapshotted in is given in the most
current snapshotted commit:

    commit fb578c9143f5489488b7adc0ddbb253eee9eed66 (github/tribits_github_snapshot, tribits_github_snapshot)
    Author: Roscoe A. Bartlett <rabartl@sandia.gov>
    Date:   Fri Jun 10 13:40:24 2022 -0600

        Automatic snapshot commit from tribits at e82c4544

        Origin repo remote tracking branch: 'github/master'
        Origin repo remote repo URL: 'github = git@github.com:TriBITSPub/TriBITS.git'

        At commit:

        commit e82c4544fbd50769b85bab45cb98ef009ca4643c
        Author:  Roscoe A. Bartlett <rabartl@sandia.gov>
        Date:    Fri Jun 10 11:08:38 2022 -0600
        Summary: Remove internal quotes for git log --pretty=format:<fmnt> strings (#485)

Later, we could sync up the FindTPLBLAS.cmake and FindTPLLAPACK.cmake modules
but that will break someone's backward compatibility.